merged-project-seoyun/backup-merged-project-20250725-100057/dags/requirements.txt
kafka-python==2.0.2
requests==2.31.0
psycopg2-binary==2.9.7
